+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.0.0a1 (unreleased)
 
-
-- Nothing changed yet.
+### Fixed
+- Unregister controlpanel NostrMetadataSyncMangementControlpanel if uninstalled
 
 
 ## 1.0.0a0 (2026-03-10)